.ai-buddy/tests/fixtures/test_project/requirements.txt
flask==2.3.0
bcrypt==4.0.1
pyjwt==2.8.0
pytest==7.4.0
python-dotenv==1.0.0